requirement.txt
djangorestframework
django-registration-redux
django-mathfilters
django-paypal
django-contact-us